Add option for when to show the buttons
move diff parts to the diffProvider
