+++ package.json
@@ -213,6 +213,16 @@
                     "type": "boolean",
                     "default": true,
                     "description": "Whether to prompt for confirmation, before submitting a saved changelist"
+                },
+                "perforce.editorButtons.diffPrevAndNext": {
+                    "type": "string",
+                    "enum": [
+                        "All diffable files",
+                        "Only on diffs",
+                        "Never"
+                    ],
+                    "default": "All diffable files",
+                    "description": "When to show buttons on the editor title menu for diffing the next / previous revision"
                 }
             }
         },
@@ -770,11 +780,21 @@
                 {
                     "command": "perforce.diffPrevious",
                     "group": "navigation@-10",
-                    "when": "perforce.currentFile.showDiffPrev"
+                    "when": "perforce.currentFile.showDiffPrev && config.perforce.editorButtons.diffPrevAndNext =~ /All/ || perforce.currentFile.showDiffPrev && config.perforce.editorButtons.diffPrevAndNext =~ /Only/ && perforce.currentFile.isPerforceOrDiff"
                 },
                 {
                     "command": "perforce.diffNext",
                     "group": "navigation@-9",
+                    "when": "perforce.currentFile.showDiffNext && config.perforce.editorButtons.diffPrevAndNext =~ /All/ || perforce.currentFile.showDiffPrev && config.perforce.editorButtons.diffPrevAndNext =~ /Only/ && perforce.currentFile.isPerforceOrDiff"
+                },
+                {
+                    "command": "perforce.diffPrevious",
+                    "group": "perforce@1",
+                    "when": "perforce.currentFile.showDiffPrev"
+                },
+                {
+                    "command": "perforce.diffNext",
+                    "group": "perforce@2",
                     "when": "perforce.currentFile.showDiffNext"
                 }
             ]
